+++ src/main/java/com/sacristan/api/TestDataInitializer.java
@@ -1,9 +1,24 @@
 package com.sacristan.api;
 
-import com.sacristan.api.global.models.*;
-import com.sacristan.api.global.models.user.*;
-import com.sacristan.api.global.models.user.extra.Role;
-import com.sacristan.api.global.repositories.*;
+import com.sacristan.api.global.entities.assignments.routineSequence.RoutineSequence;
+import com.sacristan.api.global.entities.content.category.Category;
+import com.sacristan.api.global.entities.content.category.CategoryRepository;
+import com.sacristan.api.global.entities.content.rotuine.Routine;
+import com.sacristan.api.global.entities.content.rotuine.RoutineRepository;
+import com.sacristan.api.global.entities.content.sequence.Sequence;
+import com.sacristan.api.global.entities.content.sequence.SequenceRepository;
+import com.sacristan.api.global.entities.content.step.Step;
+import com.sacristan.api.global.entities.content.weekDays.DaysOfTheWeek;
+import com.sacristan.api.global.entities.tracking.reproduction.Reproduction;
+import com.sacristan.api.global.entities.tracking.reproduction.ReproductionRepository;
+import com.sacristan.api.global.entities.tracking.status.Status;
+import com.sacristan.api.global.entities.users.student.Student;
+import com.sacristan.api.global.entities.users.student.StudentRepository;
+import com.sacristan.api.global.entities.users.teacher.Teacher;
+import com.sacristan.api.global.entities.users.teacher.TeacherRepository;
+import com.sacristan.api.global.entities.users.user.User;
+import com.sacristan.api.global.entities.users.user.UserRepository;
+import com.sacristan.api.global.entities.users.role.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
