+++ src/main/java/com/sacristan/api/global/entities/content/category/CategoryRepository.java
@@ -1,6 +1,5 @@
-package com.sacristan.api.global.repositories;
+package com.sacristan.api.global.entities.content.category;
 
-import com.sacristan.api.global.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
