+++ src/main/java/com/sacristan/api/global/entities/content/rotuine/RoutineRepository.java
@@ -1,7 +1,6 @@
-package com.sacristan.api.global.repositories;
+package com.sacristan.api.global.entities.content.rotuine;
 
-import com.sacristan.api.global.models.Category;
-import com.sacristan.api.global.models.Routine;
+import com.sacristan.api.global.entities.content.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
