+++ src/main/java/com/sacristan/api/global/entities/content/weekDays/DaysOfTheWeek.java
@@ -1,4 +1,4 @@
-package com.sacristan.api.global.models;
+package com.sacristan.api.global.entities.content.weekDays;
 
 public enum DaysOfTheWeek {
     MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY
