+++ src/main/java/com/sacristan/api/global/entities/tracking/status/Status.java
@@ -1,4 +1,4 @@
-package com.sacristan.api.global.models;
+package com.sacristan.api.global.entities.tracking.status;
 
 public enum Status {
     COMPLETED, ABANDONED, IN_PROGRESS
