+++ src/main/java/com/sacristan/api/global/entities/users/role/Role.java
@@ -1,0 +1,5 @@
+package com.sacristan.api.global.entities.users.role;
+
+public enum Role {
+    ADMIN, TEACHER, PARENT, STUDENT
+}
