--- src/main/java/com/sacristan/api/global/models/user/extra/Role.java
@@ -1,10 +1,0 @@
-package com.sacristan.api.global.models.user.extra;
-
-public enum Role {
-
-    ADMIN,
-    TEACHER,
-    PARENT,
-    STUDENT
-
-}
