+++ src/main/java/com/sacristan/api/global/specifications/StudentSpecification.java
@@ -1,7 +1,7 @@
 package com.sacristan.api.global.specifications;
 
-import com.sacristan.api.global.models.user.Student;
-import com.sacristan.api.global.models.user.User;
+import com.sacristan.api.global.entities.users.student.Student;
+import com.sacristan.api.global.entities.users.user.User;
 
 import jakarta.persistence.criteria.Join;
 import org.springframework.data.jpa.domain.PredicateSpecification;
